requirements.txt
fastapi~=0.89.1
openai~=0.26.4
python-dotenv~=0.21.1
flake8~=6.0.0
flake8-json~=21.7.0
gunicorn~=20.1.0
uvicorn~=0.20.0
asyncer~=0.0.2
